+++ R/ard_survival_survfit.R
@@ -32,13 +32,14 @@
 #' * Times should be provided using the same scale as the time variable used to fit the provided
 #'   survival fit model.
 #'
-#' @examplesIf do.call(asNamespace("cardx")$is_pkg_installed, list(pkg = c("survival", "broom"), reference_pkg = "cardx"))
+#' @examplesIf do.call(asNamespace("cardx")$is_pkg_installed, list(pkg = c("survival", "broom", "ggsurvfit"), reference_pkg = "cardx"))
 #' library(survival)
+#' library(ggsurvfit)
 #'
-#' survfit(Surv(AVAL, CNSR) ~ TRTA, cards::ADTTE) |>
+#' survfit(Surv_CNSR(AVAL, CNSR) ~ TRTA, cards::ADTTE) |>
 #'   ard_survival_survfit(times = c(60, 180))
 #'
-#' survfit(Surv(AVAL, CNSR) ~ TRTA, cards::ADTTE) |>
+#' survfit(Surv_CNSR(AVAL, CNSR) ~ TRTA, cards::ADTTE) |>
 #'   ard_survival_survfit(probs = c(0.25, 0.5, 0.75))
 #'
 #' # Competing Risks Example ---------------------------
@@ -240,7 +241,7 @@ ard_survival_survfit <- function(x, times = NULL, probs = NULL, type = NULL) {
   ) %>%
     dplyr::bind_rows() %>%
     `rownames<-`(NULL) %>%
-    dplyr::mutate(context = "survival") %>%
+    dplyr::mutate(context = "survival_survfit") %>%
     dplyr::as_tibble()
 
   if (length(x$n) == 1) df_stat <- df_stat %>% dplyr::select(-"strata")